.claude/skills/anything-to-notebooklm/feishu-read-mcp/requirements.txt
# MCP 服务器核心依赖
fastmcp>=0.2.0

# 浏览器自动化
playwright>=1.40.0

# HTTP 请求
aiohttp>=3.9.0
aiofiles>=23.0.0

# HTML 解析
beautifulsoup4>=4.12.0
lxml>=4.9.0

# 图片处理
Pillow>=10.0.0

# 其他工具
aiofiles>=23.0.0
aiohttp-client-cache>=0.6.0
